Reorganize e2e tests into subdirectories
- Move tests into logical subdirectories:
  - tests/e2e/core/ - Core functionality tests (48 passing tests)
  - tests/e2e/performance/ - Large payload performance tests
  - tests/e2e/preview/ - Style preview interaction tests
  - tests/e2e/skip/ - Broken/skipped tests needing fixes

- Remove all .skip() calls from test code
  - Tests should be in appropriate directories instead of explicitly skipped
  - Moved large payload tests to performance/
  - Moved flaky CSS theme test to skip/
  - Removed tests for features not supported by design (primitives, null)
  - Removed tests for features not yet implemented (theme picker in toolbar)

- Update package.json scripts to use directory paths
  - test:e2e now runs tests/e2e/core
  - test:e2e:perf runs tests/e2e/performance
  - test:e2e:preview runs tests/e2e/preview
  - No need to maintain explicit test file lists

- Add comprehensive README.md to each subdirectory
  - Documents purpose, characteristics, and usage
  - Explains when to run each test suite
  - Provides guidance for adding/moving tests

- Fix dist paths in moved test files
  - Updated from ../../dist to ../../../dist for new directory structure

Related to issue #4
