+++ package.json
@@ -13,9 +13,9 @@
     "test:integration": "vitest run tests/integration",
     "test:contract": "vitest run tests/contract",
     "test:update": "UPDATE_SNAPSHOTS=1 vitest run",
-    "test:e2e": "npm run build && playwright test tests/e2e/content.e2e.spec.ts tests/e2e/css-themes.spec.ts tests/e2e/fixtures.e2e.spec.ts tests/e2e/large-fixtures.e2e.spec.ts tests/e2e/options.e2e.spec.ts tests/e2e/preact-renderer.e2e.spec.ts tests/e2e/toolbar.e2e.spec.ts",
-    "test:e2e:perf": "npm run build && playwright test tests/e2e/performance.e2e.spec.ts",
-    "test:e2e:preview": "npm run build && playwright test tests/e2e/preview-*.e2e.spec.ts",
+    "test:e2e": "npm run build && playwright test tests/e2e/core",
+    "test:e2e:perf": "npm run build && playwright test tests/e2e/performance",
+    "test:e2e:preview": "npm run build && playwright test tests/e2e/preview",
     "test:e2e:all": "npm run build && playwright test tests/e2e",
     "test:perf": "node scripts/run-perf-with-output.mjs run --config vitest.config.perf.ts",
     "test:perf:smoke": "node scripts/run-perf-with-output.mjs run --config vitest.config.perf.ts tests/perf/parser-regression.smoke.spec.ts",
@@ -33,7 +33,6 @@
     "bundle:measure": "node scripts/measure-bundle.mjs",
     "style:preview:bundle": "vite build --config vite.config.preview.mjs",
     "style:preview": "node scripts/generate-style-test-page.mjs",
-    "test:e2e:preview": "playwright test tests/e2e/preview-*.e2e.spec.ts",
     "test:preview": "npm run style:preview && npm run test:e2e:preview",
     "clean": "rm -rf dist",
     "storybook": "storybook dev -p 6006",
