+++ tests/e2e/core/css-themes.spec.ts
@@ -7,7 +7,7 @@ const __dirname = dirname(fileURLToPath(import.meta.url))
 
 // ensure tokens block does not change light visuals
 test('tokens present do not change light visuals', async ({ page }) => {
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   // Load the built CSS and inject into a minimal page to avoid needing a dev server
@@ -30,7 +30,7 @@ test('tokens present do not change light visuals', async ({ page }) => {
 
 // json syntax colors parity
 test('json syntax colors match baseline in light mode', async ({ page }) => {
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   await page.setContent(`
@@ -60,7 +60,7 @@ test('json syntax colors match baseline in light mode', async ({ page }) => {
 // surfaces and buttons parity
 test('light theme parity for surfaces and buttons', async ({ page }) => {
   // Load the built CSS and inject into a minimal page to avoid needing a dev server
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   await page.setContent(`
@@ -100,7 +100,7 @@ test('light theme parity for surfaces and buttons', async ({ page }) => {
 
 // dark theme activation
 test('explicit dark via data-theme sets dark background', async ({ page }) => {
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   await page.setContent(`
@@ -125,7 +125,7 @@ test('system dark applies when not explicitly set', async ({ page }) => {
   // Enable dark mode at the browser level FIRST, before loading any content
   await page.emulateMedia({ colorScheme: 'dark' })
   
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   await page.setContent(`
@@ -152,7 +152,7 @@ test('system dark but explicit light attribute keeps light background', async ({
   // Enable dark mode at the browser level FIRST
   await page.emulateMedia({ colorScheme: 'dark' })
   
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   await page.setContent(`
@@ -174,7 +174,7 @@ test('system dark but explicit light attribute keeps light background', async ({
 
 // Simple sanity check after introducing cascade layers
 test('no visual regressions after layering', async ({ page }) => {
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   await page.setContent(`
@@ -194,7 +194,7 @@ test('no visual regressions after layering', async ({ page }) => {
 
 // accessibility focus-visible
 test('keyboard focus has visible outline (focus-visible via token)', async ({ page }) => {
-  const cssPath = resolve(__dirname, '../../dist/style.css')
+  const cssPath = resolve(__dirname, '../../../dist/style.css')
   const css = readFileSync(cssPath, 'utf8')
 
   await page.setContent(`
@@ -225,37 +225,3 @@ test('keyboard focus has visible outline (focus-visible via token)', async ({ pa
   expect(outlineColor).not.toBe('rgba(0, 0, 0, 0)')
 })
 
-// light/dark/system without styleDark.css present
-// This test involves page reload with emulateMedia which can be flaky in CI
-test.skip('STYLE_07: light and dark still render correctly without styleDark.css', async ({ page }) => {
-  const cssPath = resolve(__dirname, '../../dist/style.css')
-  const css = readFileSync(cssPath, 'utf8')
-
-  await page.setContent(`
-    <!doctype html>
-    <html>
-      <head>
-        <meta charset="utf-8" />
-        <style>${css}</style>
-      </head>
-      <body>hello</body>
-    </html>
-  `)
-
-  // light default
-  let bg = await page.evaluate(() => getComputedStyle(document.body).backgroundColor)
-  expect(bg).toBe('rgb(255, 255, 255)')
-
-  // explicit dark
-  await page.evaluate(() => { document.documentElement.dataset.theme = 'dark' })
-  bg = await page.evaluate(() => getComputedStyle(document.body).backgroundColor)
-  // Dark background is #0d1117 = rgb(13, 17, 23)
-  expect(bg).toBe('rgb(13, 17, 23)')
-
-  // system dark (no explicit)
-  await page.evaluate(() => { delete (document.documentElement as any).dataset.theme })
-  await page.emulateMedia({ colorScheme: 'dark' })
-  await page.reload()
-  bg = await page.evaluate(() => getComputedStyle(document.body).backgroundColor)
-  expect(bg).toBe('rgb(13, 17, 23)')
-})
